make a function to switch form's states
