+++ js/create-offer-element.js
@@ -4,40 +4,16 @@ import {
   Apartment,
 } from './data.js';
 
+import {
+  checkElementTextContent,
+  hideTextContent,
+  checkContentExistence,
+} from './util.js';
+
 const offersList = document.querySelector('.map__canvas');
 const offerTemplate = document.querySelector('#card').content.querySelector('.popup');
 const availableOffers = createOffers(OFFER_COUNT);
 const offerListFragment = document.createDocumentFragment();
-
-const checkElementTextContent = (offerElement, offerClasses) => {
-  for (let i = 0; i < offerClasses.length; i++) {
-    const contentElement = offerElement.querySelector(`.popup__${offerClasses[i]}`);
-    if ((!contentElement.textContent) ||
-    (contentElement.textContent === ' ₽/ночь') ||
-    (contentElement.textContent === ' комнаты для  гостей') ||
-    (contentElement.textContent === 'Заезд после , выезд до ')) {
-      contentElement.classList.add('visually-hidden');
-    }
-  }
-};
-
-const hideTextContent = (firstString, secondString) => {
-  if (firstString === '') {
-    secondString = '';
-  } else if (secondString === '') {
-    firstString = '';
-  }
-
-  return [firstString, secondString];
-};
-
-const checkContentExistence = (offerElement, offerKeys) => {
-  for (let i = 0; i < offerKeys.length; i++) {
-    if ((offerElement[offerKeys[i]] === null) || (offerElement[offerKeys[i]] === undefined)) {
-      offerElement[offerKeys[i]] = '';
-    }
-  }
-};
 
 availableOffers.forEach((availableOffer) => {
   const offerElement = offerTemplate.cloneNode(true);
