+++ js/main.js
@@ -1,1 +1,2 @@
 import './create-offer-element.js';
+import './form.js';
